refactor: simplify ProductManagement component and enhance product handling
- Removed unused props and state from ProductManagement, streamlining its functionality.
- Integrated product management hooks for better modularity and improved notification handling.
- Updated AdminPage to utilize the refactored ProductManagement, enhancing clarity and maintainability.
