+++ src/basic/components/admin/ProductManagement.tsx
@@ -2,61 +2,66 @@ import React from "react";
 import { Button } from "../ui/button/Button";
 import { ProductTable } from "../product/ProductTable";
 import { ProductForm } from "../product/ProductForm";
-import type { ProductFormState, NotificationType } from "../../types/admin";
-import type { CartItem, Product } from "../../../types";
+import type { NotificationType } from "../../types/admin";
+
+// hooks
+import { useProducts } from "../../hooks/useProducts";
+import { useCart } from "../../hooks/useCart";
+import { useProductForm } from "../../hooks/useProductForm";
 
 interface ProductManagementProps {
-  products: Product[];
-  cart: CartItem[];
-  onEditProduct: (product: Product) => void;
-  onDeleteProduct: (productId: string) => void;
-  onAddProduct: () => void;
-  showProductForm: boolean;
-  productForm: ProductFormState;
-  updateField: (field: keyof ProductFormState, value: string | number | any[]) => void;
-  editingProduct: string | null;
-  onProductSubmit: (e: React.FormEvent) => void;
-  onCancelProductForm: () => void;
   addNotification: (message: string, type: NotificationType) => void;
 }
 
-export const ProductManagement = ({
-  products,
-  cart,
-  onEditProduct,
-  onDeleteProduct,
-  onAddProduct,
-  showProductForm,
-  productForm,
-  updateField,
-  editingProduct,
-  onProductSubmit,
-  onCancelProductForm,
-  addNotification,
-}: ProductManagementProps) => {
+export const ProductManagement = ({ addNotification }: ProductManagementProps) => {
+  // NotificationType을 훅에서 기대하는 타입으로 변환
+  const handleNotification = (message: string, type?: "success" | "error" | "warning") => {
+    if (type === "success" || type === "error") {
+      addNotification(message, type);
+    }
+  };
+
+  const { products, addProduct, updateProduct, deleteProduct } = useProducts(handleNotification);
+  const { cart } = useCart(handleNotification);
+  const {
+    editingProduct,
+    productForm,
+    showProductForm,
+    updateField: updateProductField,
+    startEditProduct,
+    startAddProduct,
+    cancelProductForm,
+    handleProductSubmit,
+  } = useProductForm();
+
+  // Product Form 제출 처리
+  const handleProductFormSubmit = (e: React.FormEvent) => {
+    handleProductSubmit(e, addProduct, updateProduct);
+  };
+
   return (
     <section className="bg-white rounded-lg border border-gray-200">
       {/* 헤더 */}
       <div className="p-6 border-b border-gray-200">
         <div className="flex justify-between items-center">
           <h2 className="text-lg font-semibold">상품 목록</h2>
-          <Button onClick={onAddProduct} className="bg-gray-900 hover:bg-gray-800">
+          <Button onClick={startAddProduct} className="bg-gray-900 hover:bg-gray-800">
             새 상품 추가
           </Button>
         </div>
       </div>
 
       {/* 상품 테이블 */}
-      <ProductTable products={products} cart={cart} onEdit={onEditProduct} onDelete={onDeleteProduct} />
+      <ProductTable products={products} cart={cart} onEdit={startEditProduct} onDelete={deleteProduct} />
 
       {/* 상품 폼 */}
       {showProductForm && (
         <ProductForm
           productForm={productForm}
-          updateField={updateField}
+          updateField={updateProductField}
           editingProduct={editingProduct}
-          onSubmit={onProductSubmit}
-          onCancel={onCancelProductForm}
+          onSubmit={handleProductFormSubmit}
+          onCancel={cancelProductForm}
           addNotification={addNotification}
         />
       )}
